Update earl/resultStore; Change to folder with models.js
(Update references)
